requirements.txt
flask
dnspython
gunicorn
flask-cors
